Improvmenst to ssl cert upload (I think).
I am still unable to succeed, but this is more in the right direction I
think. Using the 3rd party library with unknown license to allow for
multipart file uploads didn't seem to work, and it was made aware to me
that there is a way to do it already in the requests module, so I
elected to do it that way.